refactor(macos-ipc): clean up ReconnectionManager and SettingsStore code by removing comments and improving clarity

+++ macos/SmoothScrollMenuBar/Sources/IPC/ReconnectionManager.swift
@@ -1,8 +1,6 @@
 import Foundation
 import os
 
-/// Manages auto-reconnection when IPC connection is lost.
-/// Uses AsyncStream from IPCClient.connectionLost — no polling needed.
 @MainActor
 final class ReconnectionManager {
     private let client: IPCClient
@@ -10,7 +8,7 @@ final class ReconnectionManager {
     private var task: Task<Void, Never>?
     private var retryCount = 0
     private let maxRetryCount = 10
-    private let baseDelay: UInt64 = 1_000_000_000 // 1s
+    private let baseDelay: UInt64 = 1_000_000_000
 
     init(client: IPCClient = .shared, settings: SettingsStore = .shared) {
         self.client = client
@@ -20,8 +18,6 @@ final class ReconnectionManager {
     func start() {
         task = Task { [weak self] in
             guard let self else { return }
-
-            // Wait for disconnect signal (no polling)
             for await _ in await self.client.connectionLost {
                 guard !Task.isCancelled else { return }
                 await self.handleDisconnect()
@@ -34,8 +30,7 @@ final class ReconnectionManager {
         settings.updateConnectionState(.reconnecting(attempt: retryCount))
 
         while retryCount < maxRetryCount && !Task.isCancelled {
-            // Exponential backoff
-            let delay = min(baseDelay * UInt64(1 << min(retryCount, 10)), 30_000_000_000)
+            let delay = min(baseDelay * UInt64(1 << min(retryCount, 10)), UInt64(30_000_000_000))
             try? await Task.sleep(nanoseconds: delay)
 
             do {
@@ -48,8 +43,6 @@ final class ReconnectionManager {
                 settings.updateConnectionState(.failed(error.localizedDescription))
             }
         }
-        
-        // Max retries reached
         retryCount = 0
     }
